Extracted test_suite_case_tiles
